feat: make the webpack env validator a local package
