+++ packages/verify-env/src/WebpackPlugin.ts
@@ -1,15 +1,14 @@
-import { verify } from "./index.ts";
-import type webpack from "webpack";
+import { verify } from "./verify.ts";
 export class VerifyEnvPlugin {
   private readonly srcPath: string;
   constructor(srcPath: string = "src/") {
     this.srcPath = srcPath;
   }
-  apply(compiler: webpack.Compiler) {
+  apply(compiler: any) {
     compiler.hooks.run.tap("VerifyEnvPlugin", () => {
       return verify(this.srcPath);
     });
   }
 }
 
-module.exports = VerifyEnvPlugin;
+module.exports = { VerifyEnvPlugin };
